Fix: error handling
The basic gist of this is adding correct error handling. whether an
error should crash the program or just the worker, it doesnt matter.
theyre all handled. the last major things to do are add a shutdown
sequence, buffered writing to `out`, and changing `String` to `&str`
so less heap allocations.