Add: debug logs, Add: little dispatch spawn delay
the debug logs are technically info logs. the libraries use the `debug!`
macro to liberally. I dont like that because it clogs up my logs, like
wtf. IMO that should be a feature flag.

the dispatch spawn delay was because i was getting rate limited because
sending 400 requests at the same time is frowned upon :(. anyway this
fixes it, im still zoom but not as zoom as i want :(.